Use Popen instead of check_call in collect_tests
`check_call` is blocking, and if the list of collected tests exceeds
the system's pipe-max-size, pushing to the pipe will also block until
data is pulled off of it. This means that the `check_call` call
will never return and hang forever.

Using `Popen` instead allows the code after the subprocess call to
execute immediately, and it will block reading until the pipe
gets closed from the other end.